refactor(http): move ErrResourceAccessDenied to http errors
