+++ api/http/errors/errors.go
@@ -4,5 +4,6 @@ import "errors"
 
 // General errors.
 var (
-	ErrUnauthorized = errors.New("Unauthorized")
+	ErrUnauthorized         = errors.New("Unauthorized")
+	ErrResourceAccessDenied = errors.New("Access denied to resource")
 )
